Refactor
Aims of this refactor:

- convert repo list to terser YAML, and get ready to introduce
  separation between eg docs and adrs within the same repo
- namespace everything
- make things a bit more OO so we can extend them (to add more metadata
  about repos, such as dependency versions) - eg GitHubRepoFetcher has
  become GitHub::Repo
- tidy up some requires etc, inline a bit

Tests all pass and site builds correctly
