+++ .gitignore
@@ -18,3 +18,4 @@
 .DS_Store
 
 Staticfile.auth
+.env
