+++ Rakefile
@@ -1,6 +1,6 @@
 require "govuk_tech_docs"
 require "rspec/core/rake_task"
-require_relative "./lib/requires"
+require_relative "./lib/teacher_services_tech_docs"
 
 RSpec::Core::RakeTask.new(:spec)
