--- lib/requires.rb
@@ -1,5 +1,0 @@
-require_relative "./http"
-require_relative "./string_to_id"
-require_relative "./external_doc"
-require_relative "./github_repo_fetcher"
-require_relative "./pages_by_category"
