+++ lib/teacher_services_tech_docs.rb
@@ -1,0 +1,22 @@
+$LOAD_PATH.unshift "./lib"
+
+# deps
+require "govuk_tech_docs"
+require "yaml"
+require "pry"
+require "octokit"
+require "faraday-http-cache"
+require "faraday_middleware"
+require "html/pipeline"
+require "uri"
+
+# utilities
+require "teacher_services_tech_docs/cached_http"
+require "teacher_services_tech_docs/pages_by_category"
+
+# business
+require "teacher_services_tech_docs/github/repo"
+require "teacher_services_tech_docs/github/markdown_file"
+
+module TeacherServicesTechDocs
+end
