+++ lib/teacher_services_tech_docs/pages_by_category.rb
@@ -1,0 +1,15 @@
+module TeacherServicesTechDocs
+  class PagesByCategory
+    attr_reader :sitemap
+
+    def initialize(sitemap)
+      @sitemap = sitemap
+    end
+
+    def pages
+      sitemap.resources
+        .select { |resource| resource.path.end_with?(".html") && resource.data.category }
+        .group_by { |resource| resource.data.category }
+    end
+  end
+end
