+++ spec/spec_helper.rb
@@ -2,4 +2,4 @@ require "webmock/rspec"
 
 require "govuk_tech_docs"
 
-require_relative "../lib/requires"
+require_relative "../lib/teacher_services_tech_docs"
